# Conflicts:
#	Assets/Prefab/CanvasPrefab/Canvas.prefab
#	Assets/Scenes/ChoiWoohyck3.unity
#	Assets/Script/Manager/InventoryManager.cs

+++ Assets/Script/Weapon/Data/WeaponData.cs
@@ -10,7 +10,7 @@ public class WeaponData
 
     public string WeaponName { get; set; }
     public int Type1 { get; }
-    public int AttackDamage { get; set; }
+    public float AttackDamage { get; set; }
     public int skillA;
     public int skillP;
     public int rType;
